# openapi.json
{
  "components": {
    "schemas": {
...
      "Collection_Action_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When this data was last verified against the source. Null where the schema keeps no fetch timestamp -- the price tables deliberately do not, since a per-row timestamp would cost gigabytes to answer a question the bar's own ts_utc already covers.",
            "title": "As Of"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Action"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` to fetch the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Collection[Action]",
        "type": "object"
      },
      "Collection_Bar_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When this data was last verified against the source. Null where the schema keeps no fetch timestamp -- the price tables deliberately do not, since a per-row timestamp would cost gigabytes to answer a question the bar's own ts_utc already covers.",
            "title": "As Of"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Bar"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` to fetch the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Collection[Bar]",
        "type": "object"
      },
...
      "Collection_FinancialFactOut_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When this data was last verified against the source. Null where the schema keeps no fetch timestamp -- the price tables deliberately do not, since a per-row timestamp would cost gigabytes to answer a question the bar's own ts_utc already covers.",
            "title": "As Of"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/FinancialFactOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` to fetch the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Collection[FinancialFactOut]",
        "type": "object"
      },
      "Collection_SymbolSummary_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When this data was last verified against the source. Null where the schema keeps no fetch timestamp -- the price tables deliberately do not, since a per-row timestamp would cost gigabytes to answer a question the bar's own ts_utc already covers.",
            "title": "As Of"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SymbolSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` to fetch the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Collection[SymbolSummary]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Resource_SymbolDetail_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "data": {
            "$ref": "#/components/schemas/SymbolDetail"
          }
        },
        "required": [
          "data"
        ],
        "title": "Resource[SymbolDetail]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "clientCredentials": {
        "flows": {
          "clientCredentials": {
            "scopes": {
              "bars:read": "Read bars data",
              "discovery:read": "Read discovery data",
              "domains:read": "Read domains data",
              "fundamentals:read": "Read fundamentals data",
              "holders:read": "Read holders data",
              "news:read": "Read news data",
              "reference:read": "Read reference data"
            },
            "tokenUrl": "/oauth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Read-only access to the yfin market data warehouse. Authenticate with the OAuth2 client credentials flow.",
    "title": "yfin Data API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/symbols": {
      "get": {
        "description": "Symbols in the universe.\n\n`active` defaults to true: an inactive row is one discovery found but\nan operator never activated, so the pipeline does not fetch it and it\nis close to empty.",
        "operationId": "list_symbols_v1_symbols_get",
        "parameters": [
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange"
            }
          },
          {
            "in": "query",
            "name": "quote_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quote Type"
            }
          },
          {
            "description": "Symbol prefix. Matches the symbol column only.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Symbol prefix. Matches the symbol column only.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection_SymbolSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "reference:read"
            ]
          }
        ],
        "summary": "List symbols",
        "tags": [
          "market"
        ]
      }
    },
    "/v1/symbols/{symbol}": {
      "get": {
        "operationId": "get_symbol_v1_symbols__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource_SymbolDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "reference:read"
            ]
          }
        ],
        "summary": "One symbol with its latest identity snapshot",
        "tags": [
          "market"
        ]
      }
    },
    "/v1/symbols/{symbol}/actions": {
      "get": {
        "operationId": "list_actions_v1_symbols__symbol__actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection_Action_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "bars:read"
            ]
          }
        ],
        "summary": "Dividends, splits and capital gains",
        "tags": [
          "market"
        ]
      }
    },
    "/v1/symbols/{symbol}/bars": {
      "get": {
        "description": "Bars for one symbol.\n\n`session` applies to intraday intervals only, and defaults to\n`regular`. That default is accident prevention: extended-hours bars\nmixed into a regular series corrupt every indicator computed from it,\ninvisibly. Above daily the flag has no meaning, so passing it there is\nrejected rather than ignored.",
        "operationId": "list_bars_v1_symbols__symbol__bars_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "regular",
                    "all"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection_Bar_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "bars:read"
            ]
          }
        ],
        "summary": "Price bars",
        "tags": [
          "market"
        ]
      }
    },
    "/v1/symbols/{symbol}/financials": {
      "get": {
        "description": "Line items for one statement.\n\n`statement` and `freq` are both required, and that is a performance\ncontract rather than a stylistic choice: `financial_facts` is keyed on\n(symbol, statement, freq, period_end, item_key), so without them the\nquery cannot use the leading columns of its own primary key.",
        "operationId": "list_financials_v1_symbols__symbol__financials_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "statement",
            "required": true,
            "schema": {
              "title": "Statement",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "freq",
            "required": true,
            "schema": {
              "title": "Freq",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection_FinancialFactOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "fundamentals:read"
            ]
          }
        ],
        "summary": "Financial statement line items",
        "tags": [
          "market"
        ]
      }
    }
  }
}